Move to api, searchable attributes
